% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascadeSelect.R
\name{oiIcons}
\alias{oiIcons}
\title{The "oi" icons}
\usage{
oiIcons()
}
\value{
A data frame.
}
\description{
The \code{oi} icons are some SVG icons. This function lists
  them. They are intended to be used in the \code{\link{Icon}} function.
}
