% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-primeicons.R
\docType{data}
\name{pi_icons}
\alias{pi_icons}
\title{The prime icons}
\format{
A vector with 260 elements. Each element is the name of an icon.
  In order to use it in the \code{\link{Icon}} function, you have to prefix
  it with \code{pi pi-} (for example \code{"pi pi-youtube"}). See
  \href{https://primereact.org/icons/#list}{PrimeIcons} for the list of
  all icons.
}
\usage{
pi_icons
}
\description{
The names of the icons available in the \strong{PrimeIcons}
  library.
}
\keyword{datasets}
