% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-themes.R
\docType{data}
\name{themes}
\alias{themes}
\title{The CSS themes}
\format{
A vector with 34 elements. Each element is the name of a theme,
  to be used as the \code{theme} argument of the
  \code{\link{cascadeSelectInput}} function.
}
\usage{
themes
}
\description{
The names of the CSS themes available in the \strong{PrimeReact}
  library.
}
\keyword{datasets}
