% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascadeSelect.R
\name{cascadeSelectInput}
\alias{cascadeSelectInput}
\title{Cascade select input}
\usage{
cascadeSelectInput(
  inputId,
  choices,
  selected = NULL,
  placeholder = "Select",
  optionLabel,
  optionGroupLabel,
  optionGroupChildren,
  theme = "bootstrap4-dark-purple"
)
}
\arguments{
\item{inputId}{the id that will be used to get the selected value in Shiny}

\item{choices}{a hierarchical list; see the examples}

\item{selected}{the selected value; \code{NULL} for none}

\item{placeholder}{placeholder appearing when no selected value}

\item{optionLabel}{the label of the options to be selected}

\item{optionGroupLabel}{the label of the groups of options}

\item{optionGroupChildren}{a list of the names of the groups of options}

\item{theme}{the CSS theme; see \code{data(themes)} for the list of
available themes}
}
\value{
A \code{shiny.tag.list} object to be included in a Shiny UI.
}
\description{
Create a cascade select input for Shiny.
}
