% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascadeSelect.R
\name{Icon}
\alias{Icon}
\title{Create an icon}
\usage{
Icon(icon = "pi pi-circle-fill", color = "red", size = "1rem")
}
\arguments{
\item{icon}{the name (actually the class name) of the icon;
\strong{PrimeIcons} are available, for example \code{"pi pi-youtube"} (see
\href{https://primereact.org/icons/#list}{PrimeIcons}),
\strong{fontawesome} icons are available, for example
\code{"fa fa-paper-plane"}, and \strong{Bootstrap icons} are available,
for example \code{"bi bi-apple"} (see
\href{https://icons.getbootstrap.com/}{Bootstrap icons})}

\item{color}{a CSS color}

\item{size}{size of the icon, a css measurement (e.g.
\code{1rem}, \code{11px})}
}
\value{
A list to be included in the fields \code{icon} of the \code{choices}
  list of \code{\link{cascadeSelectInput}}.
}
\description{
Create an icon with a given color and a given size.
}
